### 2021-08-24 #58 @CoraliePicoche
[Re] Reproductive pair correlations and the clustering of organisms
**AUTHOR**

Dear **@ReScience/editors**,

I request a review for the following replication:

### Original article

**Title**: Reproductive pair correlations and the clustering of organisms \
**Authors**: W.R. Young, A.J. Roberts & G. Stuhne\
**Journal**: Nature\
**Year**: 2001\
**DOI**: 10.1038/35085561\
**PDF**: https://www.nature.com/articles/35085561.pdf

### Replication

**Authors**: C. Picoche, W.R. Young, F. Barraquand \
**Repository**: https://github.com/CoraliePicoche/brownian_bug_fluid \
**PDF**: https://github.com/CoraliePicoche/brownian_bug_fluid/blob/main/article/article.pdf \
**Keywords**: Agent-based model ; Turbulence ; Pair correlation function ; Microscale \
**Language**: C++ / R \
**Domain**: Computational Ecology

### Results
We have fully replicated the article.

### Potential reviewers
**@pboesu** \
Cesar B. Rocha

### Suggested editor
Timothée Poisot \
Pierre de Buyl
