### 2021-09-28 #58 @CoraliePicoche
Hello,

We were wondering if there were any news regarding our submission. Has an editor been assigned yet ?

Thank you for your help.
