### 2021-11-15 #58 @FTurci
### General comments

The article reproduces the results of Young et al. Nature 2001 and provides an additional precious step-by-step derivation of the pair density function.

The authors also provide an interesting discussion of the characteristic scales for the validity of the approach, which reveals some limitations of the theoretical model and suggests possible avenues for future development.

### Code

I managed to run the code on my own machine, reproducing the simulations and visualisations of the article.

Αs a minor note, having to edit the Makefile for every figure is a bit inconvenient, so I would suggest having separate targets for different figures.

### Paper

Overall the article is very clear and is a good complement to the original paper, with a detailed derivation in the Supplementary Materials.

Minor notes

- I did not find the explicit definition of lambda and mu. Clearly, they are the death and birth rates, related to the probabilities p and q mentioned in the Methods section. For completeness, lambda, mu and gamma should be defined before or immediately after equation 3.
- "brownian" should be "Brownian"
