### 2021-11-15 #58 @rajeshrinet
@pdebuyl  I have now gone through the paper and the code. It does an excellent job at reproducing the results of Young et al Nature 2001.  I particularly appreciate the detailed derivations in the paper and the complimentary well-commented code. Both of these will be very useful for a wider audience. I have included below some minor comments on the code and the paper which the authors may consider.

## Code

I could execute the code on my system. Here are some suggestions:

* Please add a [LICENSE to the repo](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-license-to-a-repository)

* Please add a [CONTRIBUTING file](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)

* In addition, it would be great to populate the README.md on the landing page with details on what do the folders contain, etc. The authors may also consider adding a .gif file from their simulation to the README.md.

## Paper
The paper is very well-written and steps very clearly described. Minor comments on the paper follow:

* Please provide labels to the x-axis in all the panels in Fig.1. The tick sizes and axis labels in Fig.3 are small.

* I suggest adding some more citations since one of the aims of the paper is to *make this replication article more accessible to most readers*. For example, authors should cite standard texts for stochastic simulations, turbulence, Reynolds number, etc. For example, the names of Kolmogorov and Batchelor appear without any fixed citation.
