### 2021-11-16 CoraliePicoche/brownian_bug_fluid#1 @rajeshrinet
Error in plotting of Fig 3 in your submission to ReScience
@CoraliePicoche I have been having issues with plotting the Figure 3. I get the error:

>   cannot open file '../simulation/pcf_dpow0p25_area10_tmax1000_N200000_U0p0.txt': No such file or directory

Please can you check and let me know if I am missing something?

PS: I could generate all other figures and execute all the code.
