### 2021-11-16 CoraliePicoche/brownian_bug_fluid#1 @CoraliePicoche
Hello @rajeshrinet ,

No, you are not missing anything, I just need to make the code clearer.

All files "pcf_dpow0p25_area10_tmax1000_N200000_U*' can be generated by the same code, i.e. main_Fig3.cpp. However, as for now, this code only performs one simulation with one value of Utot per run. You have to change  the value of Utot (l.156) to either 0, 0.1, 0.5 or 2.5, and change the name of the files f0 (l.168) and f1 (l.170) to the corresponding extension (_U0p0 ; _U0p1 ; _U0p5 or _U2p5, respectively) to simulate the 4 situations and obtain the 4 files used in pcf_Fig3.r.
Do not hesitate to tell me if something is unclear.

I will try and explain that better in the comments of the file, thank you for telling me.

Best regards
Coralie Picoche
