### 2021-11-16 CoraliePicoche/brownian_bug_fluid#1 @rajeshrinet
@CoraliePicoche please consider adding a tests folder in your updates. You can add scripts like one [here](https://github.com/rajeshrinet/brownian_bug_fluid/blob/main/code/tests/fig1.sh) for each of the figures. This would be useful for a new user. You may also choose default parameters for system sizes and time duration such that the code runs quickly.

You can also supply actual parameters used in the paper
