### 2021-11-16 CoraliePicoche/brownian_bug_fluid#1 @CoraliePicoche
Thank you very much for the idea and the example you adapted to our case.

I am not sure about your last sentence: do you suggest to have either (or both?) a script for a small system so that new users can have a quick idea of how the program works, or to keep the actual parameters? Or do you, more simply, suggest to have a separate file providing the actual values so that users can modify the small-system script with values used in the paper? I am ok with both, I am just wondering if that is what you meant.
