### 2021-11-16 CoraliePicoche/brownian_bug_fluid#1 @rajeshrinet
@CoraliePicoche I suggest having both. I do indeed want the users to quickly see how program works and also reproduce the plots in your article. Unfortunately, I do not use c++ regularly, so I do not know what is the best way to do this. So my comment was a bit vague.

A simple way would be to have parameters in two text files. One for quick check and other for actual plots in the article. You can indicate to user both with appropriate warnings. Also give by default the parameters which run quickly in the test folder. This will be useful for any contributors the repo (and you to manage merge requests). They can quickly check if the code runs fine for a smaller system size before pushing changes to the GitHub repo.
