### 2022-03-14 #58 @CoraliePicoche
Hello @pdebuyl,

We have now corrected the manuscript and improved the clarity of the code based on both the reviews and an [issue](https://github.com/CoraliePicoche/brownian_bug_fluid/issues/1) raised by @rajeshrinet directly in the replication repository. We apologize for the delay.

Following reviewers' advice, we now give more context in the manuscript with several, beginner-oriented references. We have also corrected the formulation of some of the equations which, while still true, could be improved, and corrected a spelling mistake in equation (26).

Regarding the code, we have created a new `code/tests` directory which allows any user to run similar programs and produce corresponding figures, but with parameters adapted for faster simulations. These simulations should help the user familiarize with the functioning of the code before launching longer simulations to replicate our results. Differences between parameters are given in `code/README.md`.

We hope that our replication is now a good fit for ReScience.

Best regards,

On behalf of all co-authors,

Coralie Picoche.

++++++++++++++++++

**Reviewer 1: @FTurci**

> ### General comments
>
> The article reproduces the results of Young et al. Nature 2001 and provides an additional precious step-by-step derivation of the pair density function.
>
> The authors also provide an interesting discussion of the characteristic scales for the validity of the approach, which reveals some limitations of the theoretical model and suggests possible avenues for future development.

We thank @FTurci for his positive feedback.

> ### Code
>
> I managed to run the code on my own machine, reproducing the simulations and visualisations of the article.
>
> Αs a minor note, having to edit the Makefile for every figure is a bit inconvenient, so I would suggest having separate targets for different figures.

We have now added separate makefiles for each script.

> ### Paper
>
> Overall the article is very clear and is a good complement to the original paper, with a detailed derivation in the Supplementary Materials.
>
> Minor notes
>
>     * I did not find the explicit definition of lambda and mu. Clearly, they are the death and birth rates, related to the probabilities p and q mentioned in the Methods section. For completeness, lambda, mu and gamma should be defined before or immediately after equation 3.

We have added the definition of lambda and mu immediately after equation 3.

>     * "brownian" should be "Brownian"

We have corrected the spelling of Brownian throughout the manuscript and Supplementary.

++++++++++++++++++

**Reviewer 2: @rajeshrinet**

> @pdebuyl I have now gone through the paper and the code. It does an excellent job at reproducing the results of Young et al Nature 2001. I particularly appreciate the detailed derivations in the paper and the complimentary well-commented code. Both of these will be very useful for a wider audience. I have included below some minor comments on the code and the paper which the authors may consider.

We thank @rajeshrinet for his comments and suggestions.

> ## Code
>
> I could execute the code on my system. Here are some suggestions:
>
>     * Please add a [LICENSE to the repo](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-license-to-a-repository)

We have chosen a GPLv3 license and added the corresponding file to our repository.

>
>     * Please add a [CONTRIBUTING file](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)

We did not find many examples of such contributing file adapted to the context of Rescience (only [here](https://github.com/prateekstark/training-binary-neural-network) and [here](https://github.com/CampusAI/Hamiltonian-Generative-Networks/blob/master/CONTRIBUTING.md)). Due to the brevity of the previous examples and of our own contribution recommendation, we have chosen to keep the contribution section inside the main README.

>     * In addition, it would be great to populate the README.md on the landing page with details on what do the folders contain, etc. The authors may also consider adding a .gif file from their simulation to the README.md.

We have extended the main README: we now describe the structure of all directories as well as the main requirements to execute the code, which were previously in `code/README.md`. As mentioned above, we have also included a contribution section.

Regarding .gif files, while we think that this is a good idea to visualize the movement of particles, we fear that this could disturb the reader. We have therefore added said files in the directory and referred to them in the main README, but do not directly show them.

>
>
> ## Paper
>
> The paper is very well-written and steps very clearly described. Minor comments on the paper follow:
>
>     * Please provide labels to the x-axis in all the panels in Fig.1. The tick sizes and axis labels in Fig.3 are small.

Figures have been improved and are now in the manuscript.

>     * I suggest adding some more citations since one of the aims of the paper is to _make this replication article more accessible to most readers_. For example, authors should cite standard texts for stochastic simulations, turbulence, Reynolds number, etc. For example, the names of Kolmogorov and Batchelor appear without any fixed citation.

We have now added general references regarding turbulence (A first course in turbulence, Tennekes et al. 1972), especially applied to small organisms such as phytoplankton (Living at microscale, Dusenbery 2009) in the introduction. Both books target naive readers and are sufficient to understand the phenomena described in our model. Such context is also given in Peters and Marrasé (2000), already cited in the first version of the manuscript, where the reader can also find a definition of the Kolmogorov scale. We would however like to point out that Kolmogorov's name is used here to qualify a known physical quantity (scale relevant to characterize a flow), not to refer to a singular work we could cite.
