### 2022-04-11 #58 @pdebuyl
Hello @CoraliePicoche `main_Fig3.out` is running since more than 48h. Can you tell me how long it should take? I am running on a Xeon 2.8GHz (so, reasonable I guess) using the provided makefile.
