### 2022-04-11 #58 @CoraliePicoche
Hello @pdebuyl ,
Are you running the "tests" simulations (faster ones, in code/tests) or the original version of the code? Which value did you use for Utot?
