### 2022-04-11 #58 @pdebuyl
I run `./main_Fig3.out` from the `code/simulation` directory with the values 0.0, 0.1, 0.5, and 2.5
