### 2022-04-11 #58 @CoraliePicoche
If you are running them sequentially, 48h does not seem too bad (I must admit that this code is clearly not optimized), I would advise to wait until tomorrow. However, if you are running the different values on different processors in parallel, this seems a bit long from what I remember on my own configuration (Xeon 3.5Hz).
If you still have computing resources available, could you tell me if the test simulations for Figure 3 work properly with your configuration?
