### 2022-04-11 #58 @pdebuyl
It is running on a headless server, so it does not bother me to let it run. It runs as four independent processes though...
