### 2022-04-11 #58 @CoraliePicoche
I have checked on my own server: computation time was around 20h for a single value of Utot (again, on a slightly "better" processor).
How long does it take for the smaller simulations?
