### 2022-04-12 #58 @CoraliePicoche
Hello @pdebuyl ,
Do you have any update regarding simulation duration? Have they ended on your server?
