### 2022-04-19 #58 @pdebuyl
Hello @CoraliePicoche

I have an update but am a bit worried about the failure to run the code.

First, from `code/simulation` the code ran for 5 days without completing.

I am trying now the files in `code/tests/simulation` and the code returns a segmentation fault except if I compile with the optimization flag `-O0`.

As you and the reviewers could run the code, it likely depends on the platform, but such errors have an impact on the readers of the reproduction, who could have the same bug. The code is too slow to run without optimizations.

From the terminal output, it seems that the code spends most of its time in `PairDens` and indeed, the pair loop is slow.

Do you have any idea on how to resolve it? Ideally the code should run without the segmentation fault...

I am using g++ 10.2.1 on linux x86_64.

@FTurci @rajeshrinet could you let us know your test platform?
