### 2022-04-25 #58 @CoraliePicoche
Hello @pdebuyl,

I have tried running the code on several machines, and I think I have found the issue. A previous version of the function `PairDens` returned a double, which is not the case anymore as variables are directly modified in the function. However, the function was still coded as returning a variable. With previous versions of g++, this did not cause any issue but it seems that newer versions are more sensitive to this, leading to simulations taking much more time to run.

Here are the different configurations I used to test the code. I was able to launch full simulations (with `code/simulation/main_Fig3.cpp`) on all of these without further issue, and test simulations (with `code/tests/simulation/main_Fig3.cpp`) always ended in less than 2 minutes.

**Local computers**

For both local computers, simulations were run with `PairDens` returning no value, *i.e.*, with the modification previously mentioned.

*Laptop*

Processor: 11th Gen Intel® Core™ i7-1165G7 @ 2.80GHz \
Linux version: 5.13.0-39 [Ubuntu 20.04]\
GSL version: 2.5\
G++ version: 9.4.0\
Time to run longer simulations: 15h

*Desktop computer*

This computer was originally the one used for all simulations presented in the paper.

Processor: Intel® Xeon(R) CPU E3-1240 v5 @ 3.50GHz \
Linux version: 4.15.0-133 [Ubuntu 16.04]\
GSL version: 2.6\
G++ version: 5.4.0\
Time to run longer simulations: 21h

**Regional cluster**

I work with the 'Mésocentre de Calcul Intensif' (MCIA), the reference cluster for scientific computing in Nouvelle Aquitaine (France). The time mentioned below corresponds to the earlier version of `PairDens`, when the function was still supposed to return a double. Simulations might be shorter with `PairDens` returning nothing.

Processor: Intel® Xeon® Gold SKL-6130 @ 2,1 GHz\
Linux version: 3.10.0 [Red Hat]\
GSL version: 1.15\
G++ version: 4.8.5\
Time to run longer simulations: 44h

I was not able to reproduce the segmentation fault you obtained running file `code/tests/simulation/main_Fig3.cpp` in any of these configurations, even before correcting the `PairDens` function. Could you try and recompile/run file `code/tests/simulation/main_Fig3.cpp`, using the command line `make -f makefile_Fig3` for compilation, to see if you still obtain a seg fault or if the problem has now been solved through the correction of the `PairDens` function?

Best regards\
Coralie Picoche
