### 2022-04-26 #58 @pdebuyl
Hi thanks for the update! I could run the tests successfully now on my laptop. I'll check the longer runs as well. Is 21h for the runs the one after the other or all at once?
