### 2022-04-26 #58 @CoraliePicoche
I am glad it worked! As for the longer runs, 21h corresponds to a single simulation, but you can of course run them all at once on different processors.
