### 2022-05-03 #58 @pdebuyl
Hi @CoraliePicoche I ran all with success thanks for the fix. The figures are as expected :-) Before publishing the article, can you add the DOIs to the bib file and provide the platform information (see https://rescience.github.io/platform/) in the repo?
Else, I'd like to apologize for the delays in the editorial procedure and say that the paper is very clearly written.
