### 2022-05-03 #58 @pdebuyl
PS: adding the DOIs will provide clickable links to the references and (I guess) allow for better bibliographical referencing.
