### 2022-05-04 #58 @CoraliePicoche
Hello @pdebuyl,

I have added the DOIs to the bib file and recompiled the .pdf, the new versions are now online.

Regarding the platform information, should I mention only the main one (the "Desktop computer" mentioned above, which was used to run all simulations), or all platforms which have been used to (successfully) test the code (which would be my laptop and regional cluster, here, maybe even your own configuration)?

Thanks for you help
Coralie
