### 2022-05-04 #58 @pdebuyl
Hello, thanks for the update. Regarding the platform, only the desktop then. It is supposed to be the computer that was used to produce the figures in your pdf. The instructions are only for Python and C/C++. In your case, there should be the g++ version, with the flags, and the version of R and of the package expint.
