### 2022-05-04 #58 @CoraliePicoche
Hello @pdebuyl ,
I have added the information on the platform in code/README.md , could you confirm that this corresponds to your requirements?
If we are close to the end of the review process, should I change the "Edited by", "Received", etc. sections of the article presentation now, or will you do it? Will I be able to modify the citation recommendation at the end of the main README once a DOI has been attributed?
Best regards
Coralie
