### 2022-05-12 #58 @pdebuyl
Hello, I will edit the metadata (editor, dates, etc) for publication. Sorry for the delay, very busy week. It's been some time since I edited for ReScience so I have to check all the procedure again.
