### 2022-05-13 #58 @pdebuyl
@CoraliePicoche I still need the zenodo or software heritage archive of the repository.

Then, I'll publish the paper. The DOI will be 10.5281/zenodo.6546488 (it does not resolve yet, see procedure at https://github.com/ReScience/articles if you want more details).
