### 2022-05-13 #58 @CoraliePicoche
> @CoraliePicoche can you archive the code? There should be a zenodo or software heritage archive of the repository. I have a pull request open for the metadata.

The code is now archived and I have added the zenodo DOI to the main README. Should I also add it to the metadata.yaml?
