### 2022-05-13 #58 @pdebuyl
if you do it, there is one less "pull request" to do so it is easier I guess
