### 2022-05-13 #58 @CoraliePicoche
I added both DOI, is that ok now?
