### 2022-05-13 #58 @pdebuyl
All is ok ! The article is on zenodo https://doi.org/10.5281/zenodo.6546488 and is sent to the ReScience article repository! https://github.com/ReScience/articles/pull/26
